requirements.txt
Django==4.2.2
psycopg2==2.9.6
django-allauth==0.54.0